double division realisation

+++ calculator.cpp
@@ -3,7 +3,7 @@
 calculator::calculator(){}
  calculator::~calculator(){}
 
-int calculator::div(int first_digit , int second_digit){return first_digit / second_digit;}
+float calculator::div(int first_digit , int second_digit){return static_cast <float>(first_digit) / second_digit;}
 int calculator::sum(int first_digit , int second_digit){return first_digit + second_digit;}
 int calculator:: mult(int first_digit , int second_digit){return first_digit * second_digit;}
 int calculator::diff(int first_digit , int second_digit){return first_digit - second_digit;}
